+++ how-to/customize-workspace/client/src/framework/lifecycle.ts
@@ -7,10 +7,16 @@ import type {
 	LifecycleProviderOptions
 } from "./shapes/lifecycle-shapes";
 import type { ModuleDefinition, ModuleEntry, ModuleHelpers } from "./shapes/module-shapes";
+import { randomUUID } from "./uuid";
 
 let lifecycleModules: ModuleEntry<Lifecycle, unknown, unknown, ModuleDefinition>[] = [];
 
-const allLifecycleEvents: { [key in LifecycleEvents]?: LifecycleHandler[] } = {};
+const allLifecycleEvents: {
+	[key in LifecycleEvents]?: {
+		id: string;
+		handler: LifecycleHandler;
+	}[];
+} = {};
 
 export async function init(
 	lifecycleProviderOptions: LifecycleProviderOptions,
@@ -20,10 +26,10 @@ export async function init(
 	await initializeModules<Lifecycle>(lifecycleModules, helpers);
 
 	for (const lifecycleModule of lifecycleModules) {
-		const conditions = await lifecycleModule.implementation.get();
-		for (const condition in conditions) {
-			allLifecycleEvents[condition] = allLifecycleEvents[condition] ?? [];
-			allLifecycleEvents[condition].push(conditions[condition]);
+		const lifecycleEvents = await lifecycleModule.implementation.get();
+		for (const key in lifecycleEvents) {
+			const lifecycleEvent: LifecycleEvents = key as LifecycleEvents;
+			subscribeLifecycleEvent(lifecycleEvent, lifecycleEvents[lifecycleEvent]);
 		}
 	}
 }
@@ -33,8 +39,30 @@ export async function fireLifecycleEvent(
 	lifecycleEvent: LifecycleEvents
 ): Promise<void> {
 	if (allLifecycleEvents[lifecycleEvent]) {
-		for (const handler of allLifecycleEvents[lifecycleEvent]) {
-			await handler(platform);
+		for (const idHandler of allLifecycleEvents[lifecycleEvent]) {
+			await idHandler.handler(platform);
+		}
+	}
+}
+
+export function subscribeLifecycleEvent(
+	lifecycleEvent: LifecycleEvents,
+	lifecycleHandler: LifecycleHandler
+): string {
+	const id = randomUUID();
+	allLifecycleEvents[lifecycleEvent] = allLifecycleEvents[lifecycleEvent] ?? [];
+	allLifecycleEvents[lifecycleEvent].push({
+		id,
+		handler: lifecycleHandler
+	});
+	return id;
+}
+
+export function unsubscribeLifecycleEvent(subscriptionId: string, lifecycleEvent: LifecycleEvents): void {
+	if (allLifecycleEvents[lifecycleEvent]) {
+		const idx = allLifecycleEvents[lifecycleEvent].findIndex((l) => l.id === subscriptionId);
+		if (idx >= 0) {
+			allLifecycleEvents[lifecycleEvent].splice(idx, 1);
 		}
 	}
 }