+++ how-to/customize-workspace/client/src/framework/shapes/module-shapes.ts
@@ -1,4 +1,5 @@
 import type { CustomPaletteSet } from "@openfin/workspace/common/src/api/theming";
+import type { LifecycleEvents, LifecycleHandler } from "./lifecycle-shapes";
 import type { LoggerCreator } from "./logger-shapes";
 import type { ColorSchemeMode } from "./theme-shapes";
 
@@ -80,6 +81,20 @@ export interface ModuleHelpers {
 	 * Get the current color scheme.
 	 */
 	getCurrentColorSchemeMode(): Promise<ColorSchemeMode>;
+
+	/**
+	 * Subscribe to lifecycle events.
+	 * @param lifecycleEvent The event to subscribe to.
+	 * @param lifecycleHandler The handle for the event.
+	 */
+	subscribeLifecycleEvent(lifecycleEvent: LifecycleEvents, lifecycleHandler: LifecycleHandler): string;
+
+	/**
+	 * Unsubscribe from lifecycle events.
+	 * @param subscriptionId The id of the subscription.
+	 * @param lifecycleEvent The event to subscribe to.
+	 */
+	unsubscribeLifecycleEvent(subscriptionId: string, lifecycleEvent: LifecycleEvents): void;
 }
 
 /**